requirements.txt
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
markdown>=3.4.0
Pygments>=2.14.0
PyMuPDF>=1.23.0
pymupdf4llm>=0.3.4
dulwich>=0.20.50
keyring>=24.0.0
pyspellchecker>=0.7
